Create Crypto Bot
I created this discord bot in python to return the value of a cryptocurrency. When you type the name of the coin the bot should reply with it's current price in USD. It has 2 other commands: "-list", which returns a list with all the supported cryptocurrencies. and "-support coin_name" which returns true if the coin is found in the list and false if it isn't.